lineBot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
dj-static==0.0.6
Flask==2.3.3
frozenlist==1.4.0
future==0.18.3
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.3.0
MarkupSafe==2.1.3
multidict==6.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
six==1.16.0
static3==0.7.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
